Further reduce CMC test runtime
